:sparkles: [Feat] EventInterestField 저장 로직 추가
